% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CV.R
\name{calculate_CV}
\alias{calculate_CV}
\title{Calculating coefficient variation (CV) for every metabolite regarding QC
levels}
\usage{
calculate_CV(QC_data)
}
\arguments{
\item{QC_data}{a table with quality control samples containing columns:
`Sample type`, `Compound`  and `Value`}
}
\description{
This function calculates CV based on QC data from Biocrates.
}
\details{
This function calculates the coefficient variation value based on
the formula \eqn{100*\frac{\sqrt{var(qc)}}{\overline{qc}}}.
}
