% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{check_normality}
\alias{check_normality}
\title{Performs statistical testing for}
\usage{
check_normality(processed)
}
\arguments{
\item{processed}{}
}
\description{
This function
}
\details{
This function
}
