% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biocrates.R
\name{clean_biocrates}
\alias{clean_biocrates}
\title{Supplementary cleaning for biocrates data}
\usage{
clean_biocrates(keep_cols)
}
\arguments{
\item{keep_cols}{character vector. Indicates whether the
variables "Sample Bar Code", "Sample Description",  "Submission Name",
"Collection Date", "Species", "Material", "OP",  "Org. Info",
"Plate Production No.", "gender", "sample type", "Plate Note",
"Well Position", "Sample Volume", "Run Number", "Injection Number",
"Measurement Time" should be left as a part of clinical data. You can
provide either `all` indicating that all of the mentioned variables should
remain in the data, `none` indicating that all should be removed or a
character vector of names of columns that should remain. Note that the
variables "Sample Type" and "Sample Identification" will not be removed.}
}
\description{
Supplementary cleaning for biocrates data
}
\details{
This function returns a character vector of names of columns to be
removed from the biocrates data.
}
\keyword{internal}
