% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LOD.R
\name{complete_LOD}
\alias{complete_LOD}
\title{complete values under the limit of detection in targeted analysis}
\usage{
complete_LOD(LOD_data, LOD_table = NULL)
}
\arguments{
\item{LOD_data}{an object of metaboR_LOD_data class.}
}
\description{
This function uses the table of limit of detection for every
metabolite in the data and completes all the <LOD values with 1/2LOD from
provided table.
}
\keyword{internal}
