% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CV.R
\name{get_CV_to_remove}
\alias{get_CV_to_remove}
\title{Get metabolites with high CV}
\usage{
get_CV_to_remove(
  CV_data,
  CV_threshold = 30,
  QC_type = "QC Level 2",
  CV_table = NULL
)
}
\arguments{
\item{CV_data}{an object of metaboR_CV_data class.}

\item{CV_threshold}{a percentage threshold value of coefficient variation.
This function will return the names of metabolites CV value larger than}

\item{QC_type}{character name of QC level referential while removing compounds.
\code{CV_threshold}.}
}
\description{
This function returns metabolites which should be removed based
on CV value
}
