% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_LOD.R
\name{get_LOD_ratios}
\alias{get_LOD_ratios}
\title{Calculates <LOD ratios per column}
\usage{
get_LOD_ratios(LOD_data)
}
\arguments{
\item{LOD_data}{an object of metaboR_LOD_data class.}
}
\description{
This function creates a table with <LOD ratios ordered by value.
}
