% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_LOD.R
\name{get_sparse_columns}
\alias{get_sparse_columns}
\title{Gets columns to be removed based on <LOD proportion}
\usage{
get_sparse_columns(LOD_tbl, LOD_threshold)
}
\description{
Gets columns to be removed based on <LOD proportion
}
