% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_raw_info}
\alias{plot_raw_info}
\title{Draws plots based on raw data}
\usage{
plot_raw_info(LOD_data)
}
\arguments{
\item{LOD_data}{an object of metaboR_LOD_data class.}
}
\description{
This function plots Submission Names, `Sample Type and Gender on
three elegant bar plots.
}
\details{
This function returns an ggplot2 object.
}
