% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_biocrates}
\alias{read_biocrates}
\title{Read Biocrates data}
\usage{
read_biocrates(path)
}
\arguments{
\item{path}{aaaa}
}
\description{
Imports Biocrates data from the `.xlsx` file
}
\details{
This function uses \code{\link[readxl]{read_xlsx}}
}
