% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_clinical_data}
\alias{read_clinical_data}
\title{Reads clinical data}
\usage{
read_clinical_data(path)
}
\arguments{
\item{path}{Path to the `xlsx` file. Default NULL, denoting no clinical data
regarding considered metabolomics matrix. See details for more information.}
}
\description{
This function imports the data containing clinical information
about samples.
}
\details{
The data contained in \code{path} should have
subject id in the first column.

This function is a supplementary function for
\code{\link[metaboR]{read_data}}. It is used to load the clinical data and
associate it with the metabolomics matrix  within
\code{\link{metaboR_raw_data}} class.
}
