% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_targeted.R
\name{read_LOD_table}
\alias{read_LOD_table}
\title{Reads table containing values of limit of detection}
\usage{
read_LOD_table(path)
}
\arguments{
\item{path}{Path to the file. This file should contain a table with a column
named `Concentration`  with the values of LOD for every metabolite from the
sample.}
}
\description{
This function imports a table with LOD values.
}
\details{
This function is a supplementary function for
\code{\link[metaboR]{read_targeted}}. It is used to load the LOD table and
associate it with the experimental matrix within
\code{\link{metaboR_LOD_data}} class.
}
