% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_targeted.R
\name{read_targeted}
\alias{read_targeted}
\title{Read targeted metabolomic data}
\usage{
read_targeted(path, LOD_table, clinical_data = NULL)
}
\arguments{
\item{path}{Path to the `.xlsx` file.}

\item{LOD_table}{A table that contains the values of LOD.
\code{\link[metaboR]{read_LOD_table}}.}

\item{clinical_data}{clinical data of metaboR_clinical class. Use
\code{\link[metaboR]{metaboR_clinical}} to read clinical data.}
}
\description{
Imports targeted data from a file.
}
\details{
This function uses \code{\link[readxl]{read_xlsx}}.
}
