% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_tLC}
\alias{read_tLC}
\title{Read targeted LC data}
\usage{
read_tLC(path)
}
\arguments{
\item{path}{Path to the `xlsx` file.}
}
\description{
Imports untargeted LC data from the `.xlsx` file.
}
\details{
This function uses \code{\link[readxl]{read_xlsx}}
}
\keyword{internal}
