% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_uGC}
\alias{read_uGC}
\title{Read untargeted GC data}
\usage{
read_uGC(path)
}
\arguments{
\item{path}{Path to the `xlsx` file.}
}
\description{
Imports untargeted GC data from the `.xlsx` file.
}
\details{
This function uses \code{\link[readxl]{read_xlsx}}
}
\keyword{internal}
