% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_uLC}
\alias{read_uLC}
\title{Read untargeted LC data}
\usage{
read_uLC(path)
}
\arguments{
\item{path}{Path to the `xlsx` file.}
}
\description{
Imports untargeted LC data from the `.txt` file.
}
\details{
This function uses \code{\link[data.table]{fread}}
}
\keyword{internal}
