% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_untargeted_LC}
\alias{read_untargeted_LC}
\title{Read untargeted LC data}
\usage{
read_untargeted_LC(path, orientation)
}
\arguments{
\item{path}{Path to the `xlsx` file.}

\item{orientation}{specifies orientation of compounds - `"rowwise"` or
`"colwise"`}
}
\description{
Imports untargeted LC data from the `.txt` file.
}
\details{
This function uses \code{\link[data.table]{fread}}
}
\keyword{internal}
