% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_CV.R
\name{remove_high_CV}
\alias{remove_high_CV}
\title{Remove metabolites based on the list}
\usage{
remove_high_CV(CV_data, metabolites)
}
\arguments{
\item{CV_data}{an object of metaboR_CV_data class.}

\item{metabolites}{a character vector of names of metabolites to remove from
the data}
}
\description{
This function removes metabolites from the data based on
provided list.
}
\details{
This function returns an object of one of three classes:

- metaboR_imputation when the type of analysis is GC or LC untargeted

- metaboR_derivatives when the type of analysis is GC targeted

- metaboR_processed when the type of analysis is Lc targeted or biocrates
}
