% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_metabolites.R
\name{remove_sparse_metabolites}
\alias{remove_sparse_metabolites}
\title{Removes metabolites with provided % of < LOD values from the data}
\usage{
remove_sparse_metabolites(targeted_dat, pctg)
}
\arguments{
\item{targeted_dat}{data returned by \code{\link[metaboR]{read_biocrates}},
\code{\link[metaboR]{read_tLC}} or \code{\link[metaboR]{read_tGC}}.}

\item{pctg}{a number denoting <LOD %. The metabolites which contain more
than \code{pctg} % of  <LOD will be removed from the data.}
}
\description{
This function
}
\keyword{internal}
