% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_stats}
\alias{show_stats}
\title{Show statistics for all the columns of a data frame}
\usage{
show_stats(df, digits = 3, use_percent = F, max_levels = 6, dropna = F)
}
\description{
Show statistics for all the columns of a data frame
}
