% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_stats_by_group}
\alias{show_stats_by_group}
\title{Show statistics for all the columns of a data frame group by the group_var}
\usage{
show_stats_by_group(
  df,
  group_var,
  digits = 3,
  use_percent = F,
  max_levels = 6,
  dropna = F
)
}
\description{
Show statistics for all the columns of a data frame group by the group_var
}
