% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_stats_by_group_strata}
\alias{show_stats_by_group_strata}
\title{Show statistics for all the columns of a data frame group by group_var and stratify by strata}
\usage{
show_stats_by_group_strata(
  df,
  group_var,
  strata,
  digits = 3,
  use_percent = F,
  max_levels = 6,
  dropna = F
)
}
\description{
Show statistics for all the columns of a data frame group by group_var and stratify by strata
}
