% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_LOD.R
\name{handle_LOD}
\alias{handle_LOD}
\title{Imputing values under the limit of detection LOD in targeted metabolomics
analysis}
\usage{
handle_LOD(LOD_data)
}
\arguments{
\item{LOD_data}{an object of metaboR_LOD_data class.}
}
\description{
This function removes sparse metabolites for which the
ratio of <LOD values is large, completes <LOD values in accordance to
the provided LOD table and calculates CV of QC samples.
}
\details{
This function returns an object of class `metaboR_CV_data` which
contains metabolomics profile in long data format and `CV_table` with
coefficient variation calculated for QC samples.
}
