% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biocrates.R
\name{read_biocrates}
\alias{read_biocrates}
\title{Read Biocrates data}
\usage{
read_biocrates(path)
}
\arguments{
\item{path}{Path to the `.xlsx` file.}
}
\description{
Imports Biocrates data from the `.xlsx` file.
}
\details{
This function uses \code{\link[readxl]{read_xlsx}}.
Use \code{\link[metaboR]{metaboR_clinical}} to read clinical data.
}
