% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read metabolomics data}
\usage{
read_data(path, type, clinical_path = NULL, orientation = NULL)
}
\arguments{
\item{path}{Path to the `xlsx` file.}

\item{type}{character name of MS type. One of `"untargeted_LC"`,
`"untargeted_GC"`, `"targeted_LC"`, `"targeted_GC"` and `"biocrates"`.}

\item{orientation}{specifies orientation of compounds - "rowwise" or "colwise"}
}
\description{
Imports metabolomics data corresponding to the analysis type.
}
\details{
This function uses ...
}
\keyword{internal}
