% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_LOD.R
\name{remove_sparse_metabolites}
\alias{remove_sparse_metabolites}
\title{Removes metabolites with provided ratio of < LOD values from the data}
\usage{
remove_sparse_metabolites(LOD_data, LOD_threshold)
}
\arguments{
\item{LOD_data}{an object of metaboR_LOD_data class.}

\item{LOD_threshold}{a value from 0 to 1 denoting threshold of fraction of
LOD in metabolites from the sample. Each metabolite with  <LOD ratio
equal or greater than \code{LOD_threshold} will be removed from the data.}
}
\description{
This function removes metabolites for which the ratio of values
under the limit of detection is greater or equal to provided threshold.
}
